requirements.txt
streamlit
google-generativeai
feedparser
PyGithub
beautifulsoup4
pandas
pytz
